Implement ring buffer classes and benchmarking for performance evaluation
- Added `BaseRingBuffer` abstract class and three implementations: `RingBufferNDArray`, `RingBufferList`, and `RingBufferDeque`.
- Implemented methods for adding, retrieving, and sampling items from the buffers.
- Created a benchmarking script to evaluate the performance of each buffer implementation.
- Introduced a sampling methods benchmark to compare different sampling techniques on lists and arrays.
- Developed `NStepPER` class for prioritized experience replay with n-step transitions.
- Implemented `NumpyBuffer` for fixed-size circular storage of NumPy arrays.
- Created `RainbowDQN` class for DQN agent with prioritized experience replay and n-step learning.
- Added training and evaluation functions for reinforcement learning agents.
- Developed Atari wrapper for preprocessing and stacking frames from Atari environments.
- Created visualization tools for plotting stacked frames from Atari states.
- Implemented example scripts for training DQN agents on Atari Breakout.
- Added unit tests for `NStepPER` and `NumpyBuffer` to ensure functionality and correctness.

+++ dqn/buffers/optimized_per.py
@@ -23,7 +23,7 @@ class OptimizedPER(SimplePER):
         # Remove priorities buffer for safe
         self.priorities = None
 
-    def add(self, exp: Experience, td_error: float = 1) -> None:
+    def add(self, exp: Experience, td_error: float = 1.0) -> None:
         self.buffer.add(exp)
         priority = max(td_error, self.min_priority)
         self.sum_tree.add(priority)
@@ -38,7 +38,9 @@ class OptimizedPER(SimplePER):
         if td_errors is None:
             priorities = np.ones(batch.size, dtype=np.float32)
         else:
-            priorities = np.clip(td_errors, a_min=self.min_priority, a_max=None)
+            priorities = np.clip(
+                td_errors, a_min=self.min_priority, a_max=None, dtype=np.float32
+            )
 
         for priority in priorities:
             self.sum_tree.add(priority)
